+++ lnwallet/contextsigner/internalsigner/interface.go
